chore: move sequence batcher into sequence class
